# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Grift v2 API
  version: 2.0.0
  description: |
    受託開発会社の上流工程を AI で支援する業務基盤。
    7 つの Bounded Context に対応する REST API を提供する。

    認証: Firebase Auth ID Token (Bearer)
    テナント分離: JWT + X-Tenant-ID ヘッダーで操作対象テナントを指定
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.staging.benevolentdirector.app
    description: Staging
  - url: https://api.benevolentdirector.app
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/market-evidence:
    post:
      operationId: collectMarketEvidence
      tags: [market-evidence]
      summary: マルチプロバイダ市場調査を実行
      parameters:
        - $ref: "#/components/parameters/tenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectMarketEvidenceRequest"
      responses:
        "202":
          description: Accepted (collection job queued)
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/market-evidence/{evidenceId}:
    get:
      operationId: getMarketEvidence
      tags: [market-evidence]
      summary: 市場エビデンス詳細
      parameters:
        - $ref: "#/components/parameters/tenantIdHeader"
        - name: evidenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/AggregatedEvidence"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ─── Estimates (Estimation Context) ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
  parameters:
    tenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 操作対象のテナント（JWT claims と一致必須）
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Too Many Requests
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object

    # ─── Health ───
...
    CaseType:
      type: string
      enum: [new_project, bug_report, fix_request, feature_addition, undetermined]
...
    CollectMarketEvidenceRequest:
      type: object
      required: [case_type, context]
      properties:
        case_id:
          type: string
          format: uuid
        case_type:
          $ref: "#/components/schemas/CaseType"
        context:
          type: string
          description: プロジェクトの概要説明
          maxLength: 10000
        region:
          type: string
          default: japan
        providers:
          type: array
          items:
            type: string
            enum: [grok, brave, perplexity, gemini]
          description: 使用するプロバイダ（省略時は全プロバイダ）
    EvidenceFragment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [grok, brave, perplexity, gemini]
        hourly_rate_range:
          $ref: "#/components/schemas/Range"
        total_hours_range:
          $ref: "#/components/schemas/Range"
        team_size_range:
          $ref: "#/components/schemas/Range"
        duration_range:
          $ref: "#/components/schemas/Range"
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        provider_confidence:
          type: number
          minimum: 0
          maximum: 1
        retrieved_at:
          type: string
          format: date-time
    AggregatedEvidence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        fragments:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceFragment"
        consensus_hours_range:
          $ref: "#/components/schemas/Range"
        consensus_rate_range:
          $ref: "#/components/schemas/Range"
        overall_confidence:
          type: string
          enum: [high, medium, low]
        contradictions:
          type: array
          items:
            type: object
            properties:
              provider_a:
                type: string
              provider_b:
                type: string
              field:
                type: string
              description:
                type: string
        requires_human_review:
          type: boolean
        aggregated_at:
          type: string
          format: date-time
    Range:
      type: object
      properties:
        min:
          type: number
        max:
          type: number
    Citation:
      type: object
      properties:
        url:
          type: string
          format: uri
        title:
          type: string
        source_authority:
          type: string
          enum: [official, industry, community, unknown]
        snippet:
          type: string

    # ─── Estimate ───
...